fix: layout tweaks and component structure improvement
